+++ saleor/graphql/checkout/mutations/checkout_lines_update.py
@@ -9,12 +9,13 @@ from ....webhook.event_types import WebhookEventAsyncType
 from ...app.dataloaders import get_app_promise
 from ...checkout.types import CheckoutLine
 from ...core import ResolveInfo
-from ...core.descriptions import DEPRECATED_IN_3X_INPUT
+from ...core.descriptions import ADDED_IN_321, DEPRECATED_IN_3X_INPUT
 from ...core.doc_category import DOC_CATEGORY_CHECKOUT
 from ...core.scalars import UUID, PositiveDecimal
 from ...core.types import BaseInputObjectType, CheckoutError, NonNullList
 from ...core.utils import WebhookEventInfo
 from ...core.validators import validate_one_of_args_is_in_mutation
+from ...meta.inputs import MetadataInput, MetadataInputDescription
 from ...product.types import ProductVariant
 from ...site.dataloaders import get_site_promise
 from ...utils import ERROR_COULD_NO_RESOLVE_GLOBAL_ID
@@ -52,6 +53,15 @@ class CheckoutLineUpdateInput(BaseInputObjectType):
     )
     line_id = graphene.ID(
         description="ID of the line.",
+        required=False,
+    )
+
+    metadata = NonNullList(
+        MetadataInput,
+        description=(
+            f"Checkout line public metadata. Will add and update keys. To delete keys use deleteMetadata mutation. {ADDED_IN_321} "
+            f"{MetadataInputDescription.PUBLIC_METADATA_INPUT}"
+        ),
         required=False,
     )